QA_system/requirements.txt
typing
python-dotenv
langchain-openai
qdrant-client
langchain-community
langchain

